Create base class for CSS/HTML/RDF validators
I was starting on #128, but realised a good opportunity to do a refactor of
some of the validators.

The CSS validator could be DRY'd further but this should still consolidate
functionality.

This moves tests for rules into the base validator tests and out of the
CSS/HTML/RDF parser tests; all classes that extend the Base class implement

* Closes #83
* Closes #127
* Closes #128