Change "validators" to "scanners"
Confusingly, we name our scanners as `HTMLScanner`, `RDFScanner`,
etc. but put them in a `validators/` folder, and then have a `validator.js`
file as well with the real "validator".

This moves scanners into a folder that makes more sense so it's clear that
a scanner scans one file and collects messages for said file, but the validator
is the class that actually puts all of these pieces together and validates
an actual add-on.

*Phew*!