Initial attempt to warn about coin miners.
This does not attempt to detect all of them but primarily focuses on
detecting CoinHive for now.

There's a whole bunch to explore and block when looking at NoCoin but
I'd rather see how many false positives we get in the real world with
this code first.

Fixes #1643 but might need a few follow up issues to add more miners.