Rules for messages.json (#1801)
Fixes #1666

This adds a parser for messages.json files. The parser is called from the JSON scanner. It has a JSON schema for the messages.json files, which is currently just in the schema's directory, since it is neither imported nor a patch of an existing schema. I'm pretty sure that's not where it should be, though.

It also warns about missing placeholders and when using reserved message names (which are potentially inaccessible).

I think the naming regexp as it is documented is not enforced like this by Firefox, so it may be demoted to a warning, since the extension will still work, even with characters other than [A-Za-Z0-9_@] in message names. I don't know about placeholders, not something I have violated myself.

The errors are kind of annoying currently, since you have no position or reference to which message they happen at. I haven't really found precedent for how to handle that though.

And lastly, the documentation links should probably be shortened, so they're actually clickable.

* Rules for messages.json
* Add lowercase handling and add locale perfix for messages.json
* Add test for getLowercasePlaceholders when there are no placeholders