Configure Node policy support for renovate.
This configures renovate to always configure our nodejs related configs
to support currently active and in maintainance lts versions.

That's not necessarily how we handled this in the past (we dropped
support sometimes before maintainance ended) but I think that's a fair
strategy and the reintroduced support of node 8 will be automatically
dropped in January 2020.

https://docs.renovatebot.com/node/#configuring-support-policy