Warn about deprecated browser_style:true in MV3
Note that this does not warn about the use of options_ui or
sidebar_action, despite their browser_style defaulting to true in
Firefox 114 and earlier. This is because there are a very small number
of affected extensions, and warning about these keys without the
presence in manifest.json would be confusing to most extensions.